+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/BluetoohtSend3Fragment.kt
@@ -1,15 +1,13 @@
 package com.example.bemyplant.fragment
 
-import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentPRBinding
+import com.example.bemyplant.databinding.FragmentBS3Binding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -18,14 +16,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [pRFragment.newInstance] factory method to
+ * Use the [BluetoohtSend3Fragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class pRFragment : Fragment() {
+class BluetoohtSend3Fragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    val binding by lazy{FragmentPRBinding.inflate(layoutInflater)}
+    val binding by lazy{ FragmentBS3Binding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -38,16 +36,13 @@ class pRFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        showToast(requireContext(),"회원가입 완료")
         binding.button1.setOnClickListener {
-            findNavController().navigate(R.id.action_pRFragment_to_iSFragment2)
+            findNavController().navigate(R.id.action_bS3Fragment2_to_bRFragment)
         }
         // Inflate the layout for this fragment
         return binding.root
     }
-    private fun showToast(context: Context, message: String) {
-        Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
-    }
+
     companion object {
         /**
          * Use this factory method to create a new instance of
@@ -55,12 +50,12 @@ class pRFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment RegisterPlantFragment.
+         * @return A new instance of fragment bS3Fragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            pRFragment().apply {
+            BluetoohtSend3Fragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
