+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/BluetoothReadyFragment.kt
@@ -7,23 +7,16 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentBCBinding
+import com.example.bemyplant.databinding.FragmentBRBinding
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-/**
- * A simple [Fragment] subclass.
- * Use the [bCFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-class bCFragment : Fragment() {
-    // TODO: Rename and change types of parameters
+class BluetoothReadyFragment : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
-    val binding by lazy{FragmentBCBinding.inflate(layoutInflater)}
+    private lateinit var binding: FragmentBRBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -36,26 +29,22 @@ class bCFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding.button1.setOnClickListener {
-            findNavController().navigate(R.id.action_bCFragment2_to_bSFragment2)
-        }
-        // Inflate the layout for this fragment
+        binding = FragmentBRBinding.inflate(inflater, container, false)
         return binding.root
     }
 
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val button1 = binding.imageButton1
+        button1.setOnClickListener{
+            findNavController().navigate(R.id.action_bRFragment_to_loginFragment3)
+        }
+    }
+
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment bCFragment.
-         */
-        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            bCFragment().apply {
+            BluetoothReadyFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
