+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/BluetoothSend1Fragment.kt
@@ -5,7 +5,9 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
+import com.example.bemyplant.databinding.FragmentBSBinding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -14,14 +16,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [dRPFragment.newInstance] factory method to
+ * Use the [BluetoothSend1Fragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class dRPFragment : Fragment() {
+class BluetoothSend1Fragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-
+    val binding by lazy{FragmentBSBinding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -34,8 +36,11 @@ class dRPFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+        binding.button1.setOnClickListener {
+            findNavController().navigate(R.id.action_bSFragment2_to_bS2Fragment2)
+        }
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_d_r_p, container, false)
+        return binding.root
     }
 
     companion object {
@@ -45,12 +50,12 @@ class dRPFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment dRPFragment.
+         * @return A new instance of fragment bSFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            dRPFragment().apply {
+            BluetoothSend1Fragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
