+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/BluetoothSend2Fragment.kt
@@ -1,29 +1,36 @@
 package com.example.bemyplant.fragment
 
-import android.content.Intent
 import android.os.Bundle
+import android.os.Handler
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.os.Looper
 import androidx.navigation.fragment.findNavController
-import com.example.bemyplant.MainActivity
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentBRBinding
 
+// TODO: Rename parameter arguments, choose names that match
+// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-class bRFragment : Fragment() {
+/**
+ * A simple [Fragment] subclass.
+ * Use the [BluetoothSend2Fragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class BluetoothSend2Fragment : Fragment() {
+    // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    private lateinit var binding: FragmentBRBinding
-
+    private val delayMillis = 3000L
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
             param1 = it.getString(ARG_PARAM1)
             param2 = it.getString(ARG_PARAM2)
+
         }
     }
 
@@ -31,22 +38,30 @@ class bRFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentBRBinding.inflate(inflater, container, false)
-        return binding.root
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_bluetooth_send2, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val button1 = binding.imageButton1
-        button1.setOnClickListener{
-            findNavController().navigate(R.id.action_bRFragment_to_loginFragment3)
-        }
+        Handler(Looper.getMainLooper()).postDelayed({
+            findNavController().navigate(R.id.action_bS2Fragment2_to_bS3Fragment2)
+        },delayMillis)
     }
 
     companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment bS2Fragment.
+         */
+        // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            bRFragment().apply {
+            BluetoothSend2Fragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
