+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/BuletoothConnectFragment.kt
@@ -7,7 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentISBinding
+import com.example.bemyplant.databinding.FragmentBCBinding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -16,14 +16,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [iSFragment.newInstance] factory method to
+ * Use the [BuletoothConnectFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class iSFragment : Fragment() {
+class BuletoothConnectFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    val binding by lazy{FragmentISBinding.inflate(layoutInflater)}
+    val binding by lazy{FragmentBCBinding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -32,11 +32,14 @@ class iSFragment : Fragment() {
         }
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
-                              savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
         binding.button1.setOnClickListener {
-            findNavController().navigate(R.id.action_iSFragment2_to_sRFragment)
-        }// Inflate the layout for this fragment
+            findNavController().navigate(R.id.action_bCFragment2_to_bSFragment2)
+        }
+        // Inflate the layout for this fragment
         return binding.root
     }
 
@@ -47,15 +50,16 @@ class iSFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment imageSelectFragment.
+         * @return A new instance of fragment bCFragment.
          */
         // TODO: Rename and change types and number of parameters
-        @JvmStatic fun newInstance(param1: String, param2: String) =
-                iSFragment().apply {
-                    arguments = Bundle().apply {
-                        putString(ARG_PARAM1, param1)
-                        putString(ARG_PARAM2, param2)
-                    }
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            BuletoothConnectFragment().apply {
+                arguments = Bundle().apply {
+                    putString(ARG_PARAM1, param1)
+                    putString(ARG_PARAM2, param2)
                 }
+            }
     }
 }
