+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/ImageSelectFragment.kt
@@ -7,8 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentBCBinding
-import com.example.bemyplant.databinding.FragmentBS3Binding
+import com.example.bemyplant.databinding.FragmentISBinding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -17,14 +16,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [bS3Fragment.newInstance] factory method to
+ * Use the [ImageSelectFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class bS3Fragment : Fragment() {
+class ImageSelectFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    val binding by lazy{ FragmentBS3Binding.inflate(layoutInflater)}
+    val binding by lazy{FragmentISBinding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -33,14 +32,11 @@ class bS3Fragment : Fragment() {
         }
     }
 
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
         binding.button1.setOnClickListener {
-            findNavController().navigate(R.id.action_bS3Fragment2_to_bRFragment)
-        }
-        // Inflate the layout for this fragment
+            findNavController().navigate(R.id.action_iSFragment2_to_sRFragment)
+        }// Inflate the layout for this fragment
         return binding.root
     }
 
@@ -51,16 +47,15 @@ class bS3Fragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment bS3Fragment.
+         * @return A new instance of fragment imageSelectFragment.
          */
         // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            bS3Fragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+        @JvmStatic fun newInstance(param1: String, param2: String) =
+                ImageSelectFragment().apply {
+                    arguments = Bundle().apply {
+                        putString(ARG_PARAM1, param1)
+                        putString(ARG_PARAM2, param2)
+                    }
                 }
-            }
     }
 }
