+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/PlantRegisterFragment.kt
@@ -1,11 +1,15 @@
 package com.example.bemyplant.fragment
 
+import android.content.Context
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
+import com.example.bemyplant.databinding.FragmentPRBinding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -14,14 +18,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [bPFragment.newInstance] factory method to
+ * Use the [PlantRegisterFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class bPFragment : Fragment() {
+class PlantRegisterFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-
+    val binding by lazy{FragmentPRBinding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -34,10 +38,16 @@ class bPFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+        showToast(requireContext(),"회원가입 완료")
+        binding.button1.setOnClickListener {
+            findNavController().navigate(R.id.action_pRFragment_to_iSFragment2)
+        }
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_b_p, container, false)
+        return binding.root
     }
-
+    private fun showToast(context: Context, message: String) {
+        Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
+    }
     companion object {
         /**
          * Use this factory method to create a new instance of
@@ -45,12 +55,12 @@ class bPFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment bPFragment.
+         * @return A new instance of fragment RegisterPlantFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            bPFragment().apply {
+            PlantRegisterFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
