+++ BeMyPlant/app/src/main/java/com/example/bemyplant/fragment/SensorRegisterFragment.kt
@@ -7,7 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.bemyplant.R
-import com.example.bemyplant.databinding.FragmentBSBinding
+import com.example.bemyplant.databinding.FragmentSRBinding
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -16,14 +16,14 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [bSFragment.newInstance] factory method to
+ * Use the [SensorRegisterFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class bSFragment : Fragment() {
+class SensorRegisterFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
-    val binding by lazy{FragmentBSBinding.inflate(layoutInflater)}
+    val binding by lazy{FragmentSRBinding.inflate(layoutInflater)}
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -36,11 +36,11 @@ class bSFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding.button1.setOnClickListener {
-            findNavController().navigate(R.id.action_bSFragment2_to_bS2Fragment2)
+        binding.button4.setOnClickListener {
+            findNavController().navigate(R.id.action_sRFragment_to_bCFragment2)
         }
-        // Inflate the layout for this fragment
         return binding.root
+        // Inflate the layout for this fragment
     }
 
     companion object {
@@ -50,12 +50,12 @@ class bSFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment bSFragment.
+         * @return A new instance of fragment sRFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            bSFragment().apply {
+            SensorRegisterFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
