+++ BeMyPlant/app/src/main/res/layout/fragment_bluetooth_connect.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.bCFragment">
+    tools:context=".fragment.BuletoothConnectFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
