+++ BeMyPlant/app/src/main/res/layout/fragment_bluetooth_send1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.bSFragment">
+    tools:context=".fragment.BluetoothSend1Fragment">
 
     <!-- TODO: Update blank fragment layout -->
