+++ BeMyPlant/app/src/main/res/layout/fragment_image_select.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.iSFragment">
+    tools:context=".fragment.ImageSelectFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <LinearLayout
