+++ BeMyPlant/app/src/main/res/layout/fragment_plant_register.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.pRFragment">
+    tools:context=".fragment.PlantRegisterFragment">
 
         <TextView
             android:id="@+id/textView1"
