+++ BeMyPlant/app/src/main/res/layout/fragment_sensor_reigster.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.sRFragment">
+    tools:context=".fragment.SensorRegisterFragment">
 
     <!-- TODO: Update blank fragment layout -->
